Fix macos and windows build

+++ src/engine/client/serverbrowser_http.cpp
@@ -157,7 +157,7 @@ bool CServerBrowserHttp::Parse(json_value *pJson, std::vector<CServerInfo> *pvSe
 			continue;
 		}
 		CServerInfo SetInfo = ParsedInfo;
-		str_format(SetInfo.m_aAddress, sizeof(SetInfo.m_aAddress), "%s:%ld", Addresses.u.string.ptr, Port.u.integer);
+		str_format(SetInfo.m_aAddress, sizeof(SetInfo.m_aAddress), "%s:%d", Addresses.u.string.ptr, static_cast<int>(Port.u.integer));
 		net_addr_from_str(&SetInfo.m_NetAddr, SetInfo.m_aAddress);
 		SetInfo.m_InfoGotByHttp = true;
 		vServers.push_back(SetInfo);
